chore: refactor: extract ClusterConfig
Extract ClusterConfig and related types.
Make one huge file a bit smaller.
