+++ pkg/machinery/config/types/v1alpha1/v1alpha1_provider.go
@@ -11,16 +11,13 @@ import (
 	"fmt"
 	"log"
 	"net"
-	"net/url"
 	"os"
-	goruntime "runtime"
 	"strings"
 	"time"
 
 	specs "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/talos-systems/crypto/x509"
 	"github.com/talos-systems/go-blockdevice/blockdevice/util/disk"
-	talosnet "github.com/talos-systems/net"
 
 	"github.com/talos-systems/talos/pkg/machinery/config"
 	"github.com/talos-systems/talos/pkg/machinery/config/encoder"
@@ -309,251 +306,6 @@ func (k *KubeletConfig) RegisterWithFQDN() bool {
 	return k.KubeletRegisterWithFQDN
 }
 
-// Name implements the config.Provider interface.
-func (c *ClusterConfig) Name() string {
-	return c.ClusterName
-}
-
-// Endpoint implements the config.Provider interface.
-func (c *ClusterConfig) Endpoint() *url.URL {
-	return c.ControlPlane.Endpoint.URL
-}
-
-// LocalAPIServerPort implements the config.Provider interface.
-func (c *ClusterConfig) LocalAPIServerPort() int {
-	if c.ControlPlane.LocalAPIServerPort == 0 {
-		return constants.DefaultControlPlanePort
-	}
-
-	return c.ControlPlane.LocalAPIServerPort
-}
-
-// CertSANs implements the config.Provider interface.
-func (c *ClusterConfig) CertSANs() []string {
-	return c.APIServerConfig.CertSANs
-}
-
-// CA implements the config.Provider interface.
-func (c *ClusterConfig) CA() *x509.PEMEncodedCertificateAndKey {
-	return c.ClusterCA
-}
-
-// AggregatorCA implements the config.Provider interface.
-func (c *ClusterConfig) AggregatorCA() *x509.PEMEncodedCertificateAndKey {
-	return c.ClusterAggregatorCA
-}
-
-// ServiceAccount implements the config.Provider interface.
-func (c *ClusterConfig) ServiceAccount() *x509.PEMEncodedKey {
-	return c.ClusterServiceAccount
-}
-
-// AESCBCEncryptionSecret implements the config.Provider interface.
-func (c *ClusterConfig) AESCBCEncryptionSecret() string {
-	return c.ClusterAESCBCEncryptionSecret
-}
-
-// Config implements the config.Provider interface.
-func (c *ClusterConfig) Config(t machine.Type) (string, error) {
-	return "", nil
-}
-
-// APIServer implements the config.Provider interface.
-func (c *ClusterConfig) APIServer() config.APIServer {
-	if c.APIServerConfig == nil {
-		return &APIServerConfig{}
-	}
-
-	return c.APIServerConfig
-}
-
-// Image implements the config.Provider interface.
-func (a *APIServerConfig) Image() string {
-	image := a.ContainerImage
-
-	if image == "" {
-		image = fmt.Sprintf("%s:v%s", constants.KubernetesAPIServerImage, constants.DefaultKubernetesVersion)
-	}
-
-	return image
-}
-
-// ExtraArgs implements the config.Provider interface.
-func (a *APIServerConfig) ExtraArgs() map[string]string {
-	return a.ExtraArgsConfig
-}
-
-// ExtraVolumes implements the config.Provider interface.
-func (a *APIServerConfig) ExtraVolumes() []config.VolumeMount {
-	volumes := make([]config.VolumeMount, 0, len(a.ExtraVolumesConfig))
-
-	for _, volume := range a.ExtraVolumesConfig {
-		volumes = append(volumes, volume)
-	}
-
-	return volumes
-}
-
-// ControllerManager implements the config.Provider interface.
-func (c *ClusterConfig) ControllerManager() config.ControllerManager {
-	if c.ControllerManagerConfig == nil {
-		return &ControllerManagerConfig{}
-	}
-
-	return c.ControllerManagerConfig
-}
-
-// Image implements the config.Provider interface.
-func (c *ControllerManagerConfig) Image() string {
-	image := c.ContainerImage
-
-	if image == "" {
-		image = fmt.Sprintf("%s:v%s", constants.KubernetesControllerManagerImage, constants.DefaultKubernetesVersion)
-	}
-
-	return image
-}
-
-// ExtraArgs implements the config.Provider interface.
-func (c *ControllerManagerConfig) ExtraArgs() map[string]string {
-	return c.ExtraArgsConfig
-}
-
-// ExtraVolumes implements the config.Provider interface.
-func (c *ControllerManagerConfig) ExtraVolumes() []config.VolumeMount {
-	volumes := make([]config.VolumeMount, 0, len(c.ExtraVolumesConfig))
-
-	for _, volume := range c.ExtraVolumesConfig {
-		volumes = append(volumes, volume)
-	}
-
-	return volumes
-}
-
-// Proxy implements the config.Provider interface.
-func (c *ClusterConfig) Proxy() config.Proxy {
-	if c.ProxyConfig == nil {
-		return &ProxyConfig{}
-	}
-
-	return c.ProxyConfig
-}
-
-// Enabled implements the config.Provider interface.
-func (p *ProxyConfig) Enabled() bool {
-	return !p.Disabled
-}
-
-// Image implements the config.Provider interface.
-func (p *ProxyConfig) Image() string {
-	image := p.ContainerImage
-
-	if image == "" {
-		image = fmt.Sprintf("%s:v%s", constants.KubeProxyImage, constants.DefaultKubernetesVersion)
-	}
-
-	return image
-}
-
-// Mode implements the Proxy interface.
-func (p *ProxyConfig) Mode() string {
-	if p.ModeConfig == "" {
-		return "iptables"
-	}
-
-	return p.ModeConfig
-}
-
-// ExtraArgs implements the Proxy interface.
-func (p *ProxyConfig) ExtraArgs() map[string]string {
-	return p.ExtraArgsConfig
-}
-
-// Scheduler implements the config.Provider interface.
-func (c *ClusterConfig) Scheduler() config.Scheduler {
-	if c.SchedulerConfig == nil {
-		return &SchedulerConfig{}
-	}
-
-	return c.SchedulerConfig
-}
-
-// AdminKubeconfig implements the config.Provider interface.
-func (c *ClusterConfig) AdminKubeconfig() config.AdminKubeconfig {
-	return c.AdminKubeconfigConfig
-}
-
-// ScheduleOnMasters implements the config.Provider interface.
-func (c *ClusterConfig) ScheduleOnMasters() bool {
-	return c.AllowSchedulingOnMasters
-}
-
-// Image implements the config.Provider interface.
-func (s *SchedulerConfig) Image() string {
-	image := s.ContainerImage
-
-	if image == "" {
-		image = fmt.Sprintf("%s:v%s", constants.KubernetesSchedulerImage, constants.DefaultKubernetesVersion)
-	}
-
-	return image
-}
-
-// ExtraArgs implements the config.Provider interface.
-func (s *SchedulerConfig) ExtraArgs() map[string]string {
-	return s.ExtraArgsConfig
-}
-
-// ExtraVolumes implements the config.Provider interface.
-func (s *SchedulerConfig) ExtraVolumes() []config.VolumeMount {
-	volumes := make([]config.VolumeMount, 0, len(s.ExtraVolumesConfig))
-
-	for _, volume := range s.ExtraVolumesConfig {
-		volumes = append(volumes, volume)
-	}
-
-	return volumes
-}
-
-// Etcd implements the config.Provider interface.
-func (c *ClusterConfig) Etcd() config.Etcd {
-	if c.EtcdConfig == nil {
-		c.EtcdConfig = &EtcdConfig{}
-	}
-
-	return c.EtcdConfig
-}
-
-// Image implements the config.Provider interface.
-func (e *EtcdConfig) Image() string {
-	image := e.ContainerImage
-	suffix := ""
-
-	if goruntime.GOARCH == "arm64" {
-		suffix = "-arm64"
-	}
-
-	if image == "" {
-		image = fmt.Sprintf("%s:%s%s", constants.EtcdImage, constants.DefaultEtcdVersion, suffix)
-	}
-
-	return image
-}
-
-// CA implements the config.Provider interface.
-func (e *EtcdConfig) CA() *x509.PEMEncodedCertificateAndKey {
-	return e.RootCA
-}
-
-// ExtraArgs implements the config.Provider interface.
-func (e *EtcdConfig) ExtraArgs() map[string]string {
-	if e.EtcdExtraArgs == nil {
-		e.EtcdExtraArgs = make(map[string]string)
-	}
-
-	return e.EtcdExtraArgs
-}
-
 // Mirrors implements the Registries interface.
 func (r *RegistriesConfig) Mirrors() map[string]config.RegistryMirrorConfig {
 	mirrors := make(map[string]config.RegistryMirrorConfig, len(r.RegistryMirrors))
@@ -652,123 +404,6 @@ func (r *RegistryTLSConfig) GetTLSConfig() (*tls.Config, error) {
 	}
 
 	return tlsConfig, nil
-}
-
-// Token implements the config.Provider interface.
-func (c *ClusterConfig) Token() config.Token {
-	return c
-}
-
-// ID implements the config.Provider interface.
-func (c *ClusterConfig) ID() string {
-	parts := strings.Split(c.BootstrapToken, ".")
-	if len(parts) != 2 {
-		return ""
-	}
-
-	return parts[0]
-}
-
-// Secret implements the config.Provider interface.
-func (c *ClusterConfig) Secret() string {
-	parts := strings.Split(c.BootstrapToken, ".")
-	if len(parts) != 2 {
-		return ""
-	}
-
-	return parts[1]
-}
-
-// Network implements the config.Provider interface.
-func (c *ClusterConfig) Network() config.ClusterNetwork {
-	return c
-}
-
-// DNSDomain implements the config.Provider interface.
-func (c *ClusterConfig) DNSDomain() string {
-	if c.ClusterNetwork == nil {
-		return constants.DefaultDNSDomain
-	}
-
-	return c.ClusterNetwork.DNSDomain
-}
-
-// CNI implements the config.Provider interface.
-func (c *ClusterConfig) CNI() config.CNI {
-	switch {
-	case c.ClusterNetwork == nil:
-		fallthrough
-
-	case c.ClusterNetwork.CNI == nil:
-		return &CNIConfig{
-			CNIName: constants.DefaultCNI,
-		}
-	}
-
-	return c.ClusterNetwork.CNI
-}
-
-// PodCIDR implements the config.Provider interface.
-func (c *ClusterConfig) PodCIDR() string {
-	switch {
-	case c.ClusterNetwork == nil:
-		fallthrough
-	case len(c.ClusterNetwork.PodSubnet) == 0:
-		return constants.DefaultIPv4PodNet
-	}
-
-	return strings.Join(c.ClusterNetwork.PodSubnet, ",")
-}
-
-// ServiceCIDR implements the config.Provider interface.
-func (c *ClusterConfig) ServiceCIDR() string {
-	switch {
-	case c.ClusterNetwork == nil:
-		fallthrough
-	case len(c.ClusterNetwork.ServiceSubnet) == 0:
-		return constants.DefaultIPv4ServiceNet
-	}
-
-	return strings.Join(c.ClusterNetwork.ServiceSubnet, ",")
-}
-
-// APIServerIPs returns kube-apiserver IPs in the ServiceCIDR.
-func (c *ClusterConfig) APIServerIPs() ([]net.IP, error) {
-	serviceCIDRs, err := talosnet.SplitCIDRs(c.ServiceCIDR())
-	if err != nil {
-		return nil, fmt.Errorf("failed to process Service CIDRs: %w", err)
-	}
-
-	return talosnet.NthIPInCIDRSet(serviceCIDRs, 1)
-}
-
-// DNSServiceIPs returns DNS service IPs in the ServiceCIDR.
-func (c *ClusterConfig) DNSServiceIPs() ([]net.IP, error) {
-	serviceCIDRs, err := talosnet.SplitCIDRs(c.ServiceCIDR())
-	if err != nil {
-		return nil, fmt.Errorf("failed to process Service CIDRs: %w", err)
-	}
-
-	return talosnet.NthIPInCIDRSet(serviceCIDRs, 10)
-}
-
-// ExtraManifestURLs implements the config.Provider interface.
-func (c *ClusterConfig) ExtraManifestURLs() []string {
-	return c.ExtraManifests
-}
-
-// ExtraManifestHeaderMap implements the config.Provider interface.
-func (c *ClusterConfig) ExtraManifestHeaderMap() map[string]string {
-	return c.ExtraManifestHeaders
-}
-
-// CoreDNS implements the config.Provider interface.
-func (c *ClusterConfig) CoreDNS() config.CoreDNS {
-	if c.CoreDNSConfig == nil {
-		return &CoreDNS{}
-	}
-
-	return c.CoreDNSConfig
 }
 
 // Name implements the config.Provider interface.
