+++ pkg/machinery/config/types/v1alpha1/v1alpha1_provider_test.go
@@ -9,8 +9,10 @@ import (
 	"net"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/talos-systems/talos/pkg/machinery/config"
 	"github.com/talos-systems/talos/pkg/machinery/config/types/v1alpha1"
 	"github.com/talos-systems/talos/pkg/machinery/config/types/v1alpha1/bundle"
 	"github.com/talos-systems/talos/pkg/machinery/constants"
@@ -91,4 +93,14 @@ func TestApplyDynamicConfig(t *testing.T) {
 	err = config.ApplyDynamicConfig(ctx, provider)
 	require.NoError(t, err)
 	require.Equal(t, []string{"10.2.0.3", "10.10.1.2", "10.10.1.3", "192.168.88.77"}, c.MachineConfig.CertSANs())
+}
+
+func TestInterfaces(t *testing.T) {
+	t.Parallel()
+
+	assert.Implements(t, (*config.APIServer)(nil), (*v1alpha1.APIServerConfig)(nil))
+	assert.Implements(t, (*config.ClusterConfig)(nil), (*v1alpha1.ClusterConfig)(nil))
+	assert.Implements(t, (*config.ControllerManager)(nil), (*v1alpha1.ControllerManagerConfig)(nil))
+	assert.Implements(t, (*config.Etcd)(nil), (*v1alpha1.EtcdConfig)(nil))
+	assert.Implements(t, (*config.Scheduler)(nil), (*v1alpha1.SchedulerConfig)(nil))
 }
